### 2025-06-06 #8364 @danielskatz
👋 @DjentleViBe - note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html) and note that you can click on the error above to find out more about it. I note that your date is not in the right format, which might be only issue.  Please feel free to make changes to your .md file, then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Once this works, please run `@editorialbot check references` to check the references, and then if there are any missing or malformed DOIs, please fix them.
