### 2025-06-09 #8364 @matthewfeickert
> [@matthewfeickert](https://github.com/matthewfeickert) - Would you be able to edit this JOSS submission?

@danielskatz Thanks for thinking of me for this. I think I could, though I will realistically slower to act on some reviews over the next month as I'm heading into a month of continuous travel and tight deadlines for work.

However, before I would want to spend reviewer time on this I'd want to see [documentation](https://joss.readthedocs.io/en/latest/review_criteria.html#documentation) added to the project. At the moment there's no documentation website, and the ["documentation" section of the README](https://github.com/DjentleViBe/PIRO/tree/50308aa44fca62104da6ecdfa58fb2c8f6fb64ff?tab=readme-ov-file#documentation) only links to some sparse Markdown files on `.ini` configuration files. There's also no robust test suite from first look.

Without documentation there's not much for reviewers to go off, so I think that would be important for @DjentleViBe to address before starting a review.
