### 2025-06-10 a4a3745
removing redundant .md files
