+++ BUILD.md
@@ -1,11 +1,0 @@
-## Project compilation requirements
-Requirement listed in the following table have been tested.
-
-| Category                    | Sub-category          |MacOS                              | Windows          | Linux                     |
-|-----------------------------|-----------------------|-----------------------------------|------------------|---------------------------|
-| **OS type and version**     | Name and architecture |macOS Sequoia 15.5 x64             | C++              | Arch Linux 2025.05.01, kernel 6.14.4
-| **Language and compiler**   | Language              |C++                                | C++              | C++         
-|                             | Standard              |C++23                              | C++23            | C++23       
-|                             | Compiler              |Apple clang version 17.0.0         | g++ (GCC) 11.4.0 | g++ (GCC) 15.1.1 20250425       
-| **Build tools**             | make                  |GNUmake 3.81                       | GNU Make 4.4.1         | GNU Make 4.4.1
-| **External dependencies**   | frameworks & libraries|Corefoundation, Accelerate, OpenCL | gitbash, cygwin          | opencl-nvidia 570.144-3
