+++ dependencies/assets/BC.md
@@ -1,2 +1,0 @@
-# BC
-This section is used to give details on various fields are parameters used in ```.ini``` files for Initial Condition under ```/dependencies/assets/BC```.
