+++ README.md
@@ -26,10 +26,16 @@
 - [License](#license)
 
 ## About 
-**PIRO** is an open-source C++/OpenCL software for performing high-performance numerical simulations. It provides modular solvers, supports various sparse formats (CSR, COO), and offers kernel-level customization for advanced users. PIRO is designed for research and production workflows requiring scalable parallel matrix computations.
+**PIRO** is an open-source C++/OpenCL software for performing high-performance numerical simulations. It provides modular solvers, supports various sparse formats (CSR, COO, HT), and offers kernel-level customization for advanced users. PIRO is designed for research and production workflows requiring scalable parallel matrix computations.
 <p align="center">
     <img width="250px" src="images/output.gif" align="center" alt="PIRO Advection" />
 </p>
+
+## Installation instructions
+See the [Quick installation](https://github.com/DjentleViBe/PIRO/wiki/end-to-end-example) section for step-by-step guidance.
+
+## Full Documentation
+See the [Wiki](https://github.com/DjentleViBe/PIRO/wiki) for full documentation, examples, operational details and other information.
 
 ## Features
 - __Cross-platform CPU / GPU operation__ via OpenCL.
@@ -45,7 +51,7 @@
 | $\frac{d}{dx}$ | First order spatial derivative | `solver.gradient("U")` | `Release` |```CSR```
 | $\vec{v}$ . $\nabla$ | Directional derivative along $\vec{v}$  | `solver.vector("v") & solver.gradient("U")` | `Release` | ```CSR```
 ***
-Refer to [setup.md](./dependencies/assets/setup.md) for a list of available precompiled equations
+
 ### Schemes
 | Name | Description | Number | Status |
 | --- | --- | --- | --- |
@@ -61,95 +67,6 @@ Refer to [setup.md](./dependencies/assets/setup.md) for a list of available prec
 | LU Decomposition | Solver | 27 | `Dev` |
 ***
 <p align="right">(<a href="#readme-top">back to top</a>)</p>
-
-## Running
-### Requirements
-- git
-
-If you're using Windows, run the commands in [Git Bash](https://git-scm.com/downloads).
-### Instructions
-1. To avoid Git changing line endings when you clone or commit, run:
-```bash
-git config --global core.autocrlf false
-```
-2. Clone the repo:
-```bash
-git clone https://github.com/DjentleViBe/PIRO.git
-```
-3. Download the ```.zip``` file for your Operating System from the latest [release](https://github.com/DjentleViBe/PIRO/releases/new). Extract and copy the ```bin``` folder to the project directory.
-4. From the project root directory, execute the command (Refer [Building](#how-to-build) instructions for compiling the program on your own):
-```bash
-./bin/<program name>
-``` 
-
-| program name              | Operating Sytem |
-| --- | --- |
-|```Piro_devices_MACOS```   | MacOS (x64) |
-|```Piro_devices_WIN.exe``` | Windows(x64)|
-|```Piro_devices_LIN```     | Linux (x64) |
-
-Device details are printed to the console. 
-
-5. Make the required changes to the setup file as outlined in [setup.md](./dependencies/assets/setup.md). Make sure all the ```.ini``` files are in the LF format (they should be if Step 1. was successful). ```setup.ini``` files for precompiled equations are provided in the Examples folder inside the ```.zip``` file downloaded in Step 3.
-
-6. Execute the run command ```./runall.sh -f setup.ini 0``` to initiate the simulation. Log file is saved in the [logs](./logs) folder, using the same name as the setup file.
-7. The results are saved in the [bin](./bin) directory, within a folder named after the casename parameter specified in the setup file.
-
[…]
