+++ Examples/AdvEq.ini
@@ -1,0 +1,87 @@
+[File]
+casename = AdvEq
+restart = 1
+simindex = 1
+
+[Device]
+id = 2
+platformid = 0
+;0 = CPU serial
+;1 = CPU parallel
+;2 = GPU parallel
+type = 2
+
+[Mesh]
+MeshType = 0
+; o = center location
+o = 0 0 0
+; n = number of cells in X, Y and Z
+n = 31 31 31
+; l = length in x , y and z directions
+l = 31 31 31
+; scaling factor in x, y, z directions
+s = 1 1 1
+levels = 1
+index = 1
+
+[Simulation]
+Constants = Alpha
+Values = 0.1
+Scalars = U
+Vectors = v
+
+[IC]
+; for all scalars and vectors in order
+; 0 = set all values from values section
+; 1 = set from config file
+; 2 = set from csv file
+; Types = Coulomb, Gaussian, UniformVector
+type = 1 1
+filename = Gaussian UniformVector
+
+[BC]
+; facestype - 0, 1, 2, 3 ,4, 5 ,6, 7
+; bctype - 0 : Dirichlet 1 : Neumann 2: Robin 3: Mixed 4: Cauchy
+; bc - YZ, YZ + l, XZ + b, XZ, XY, XY + h
+type = 0 0 0 0 0 0
+property = U U U U U U
+values = 0.0 0.0 0.0 0.0 0.0 0.0
+
+[Schemes]
+;1 : 1st order upwind
+;2 : 1st order downwind
+;3 : 1st order central difference
+;4 : Lax-Wendroff method
+;11 : Forward Euler
+;12 : Backward Euler
+;13 : Crank-Nicholson
+;14 : Leapfrog
+;15 : RK2
+;16 : RK4
+;27 : Implicit LU decomposition
+;28 : QR decomposition
+;37 : GMRES
+;38 : BiCGSTAB
+;39 : CG
+;40 : JLI
+Time = 11
+Space = 1
+Solver = 27
+
+[Solve]
+;0 : CSR
+;1 : Dense
+;2 : Hashtables
+;3 : COO
+Data = 0
+Timestep = 0.001
+TotalTime = 0.1
+Iterations = 0
+Save = 1
+
+[PostProcess]
+RunICS = 0
+
+[Debug]
+; 0 = INFO, 1 = DEBUG, 2 = WARNING, 3 = ERROR, 4 = TRACE
+Verbose = 1
