### 2025-06-11 cf275d5
LU decomp now goes through for testing
