### 2025-06-12 ba83f85
removed kernel compilation
