### 2025-06-12 3892056
added kernel compilation
