+++ makerun.sh
@@ -1,3 +1,4 @@
 #./makerun.sh
 ./makeall.sh
+./runall.sh -f setup.ini 1
 ./runall.sh -f setup.ini 0
