### 2025-06-12 8f786e2
modifes setup.ini in bin folder directly
