### 2025-06-12 e4e4ded
linux req script
