### 2025-06-13 ef5effc
added supprot for amd and intel
