### 2025-06-15 1ea15dd
creates .zip for executables
