### 2025-06-15 bd34bec
kernel build transfereed to makeall script
