### 2025-06-15 a4f6101
creating executable zip fies without .cl
