+++ makeall.sh
@@ -18,19 +18,48 @@ echo "Run : KERNEL SUCCESSFUL"
 mkdir -p Executables
 if [[ "$OSTYPE" == "linux-gnu"* ]]; then
     echo "Creating Linux.zip"
-    rm -f Linux.zip
-    zip -r Executables/Linux.zip bin -x '*.cl'
+    rm -f Executables/Linux.zip
+    mkdir -p Tempzip/bin
+    rsync -av \
+  --include='assets/' \
+  --include='assets/***' \
+  --include='*.bin' \
+  --include='*.exe' \
+  --exclude='*' \
+  bin/ Tempzip/bin
+    rm Tempzip/bin/assets/kernels/*.cl
+    zip -r Executables/Linux.zip Tempzip/bin
+    rm -r Tempzip
 elif [[ "$OSTYPE" == "darwin"* ]]; then
     echo "Creating MacOS.zip"
-    rm -f MacOS.zip
-    zip -r Executables/MacOS.zip bin -x '*.cl'
+    rm -f Executables
+    mkdir -p Tempzip/bin
+    rsync -av \
+  --include='assets/' \
+  --include='assets/***' \
+  --include='*.bin' \
+  --include='*.exe' \
+  --exclude='*' \
+  bin/ Tempzip/bin
+    rm Tempzip/bin/assets/kernels/*.cl
+    rm Tempzip/assets/kernels/*.cl
+    zip -r Executables/MacOS.zip Tempzip/bin
+    rm -r Tempzip
 elif [[ "$OSTYPE" == "cygwin" ]]; then
     echo "Operating System: Cygwin (POSIX compatibility layer on Windows)"
 elif [[ "$OSTYPE" == "msys" ]]; then
     echo "Creating Windows.zip"
-    mkdir -p Tempzip
-    rsync -av --exclude='*.cl' bin/ Tempzip/
-    powershell.exe Compress-Archive -Path '"Tempzip"' -DestinationPath '"Executables\\Windows.zip"' -Force
+    mkdir -p Tempzip/bin
+    rsync -av \
+  --include='assets/' \
+  --include='assets/***' \
+  --include='*.bin' \
+  --include='*.exe' \
+  --exclude='*' \
+  bin/ Tempzip/bin
+    rm Tempzip/bin/assets/kernels/*.cl
+    powershell.exe Compress-Archive -Path '"Tempzip/bin"' -DestinationPath '"Executables\\Windows.zip"' -Force
+    rm -r Tempzip
 elif [[ "$OSTYPE" == "win32" ]]; then
     echo "Operating System: Windows"
 elif [[ "$OSTYPE" == "freebsd"* ]]; then
