### 2025-06-15 1d71023
zip fole logic for mac and linux
