### 2025-06-15 1219608
rsync for arch
