### 2025-07-02 #8364 @matthewfeickert
@DjentleViBe Sorry, your response got missed while I was traveling. Let me review this update tomorrow, though given

> I thought about setting up a GitHub actions sequence for test but since the software is hardware dependent, running the tests in a CI environment isn't straightforward.

how are you planning for reviewers to evaluate the code if you aren't able to run it in CI? I understand that you can't run GPU operations in CI, but the environment installation should be something documented enough that any reviewer should be able to step through which seems like it should be able to be captured in a CI workflow.
