### 2025-07-02 #8364 @DjentleViBe
@matthewfeickert

> how are you planning for reviewers to evaluate the code if you aren't able to run it in CI? I understand that you can't run GPU operations in CI, but the environment installation should be something documented enough that any reviewer should be able to step through which seems like it should be able to be captured in a CI workflow.

Documentation contains [scripts](https://github.com/DjentleViBe/PIRO/wiki/Requirements) which handle the environment installation based on OS. Once the pre-requisites are met, one can head over to [Testing](https://github.com/DjentleViBe/PIRO/wiki/testing) and follow through.
It's just that all the scripts would have to be run locally or in a docker with supporting hardware.
