### 2025-08-04 #8364 @matthewfeickert
> what's going on with this submission?

@danielskatz Currently still trying to find reviewers. I have been getting a lot of "no"s from people given time constraints. In addition to repinging everyone I haven't gotten a reply from, this weekend I've sent 6 new requests.

@DjentleViBe, while it isn't your responsibility to find reviewers, if you have suggestions of people that you think would be good reviewers that you don't have a current [conflict of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) with please reply here with their names (please do _not_ tag them using their GitHub handle. It is fine to reference them, but don't tag them. e.g. `matthewfeickert` ✔️, @matthewfeickert ❌).
