### 2025-08-05 #8750 @editorialbot
[REVIEW]: A Sparse Matrix parallel solver for OpenCL
**Submitting author:** @DjentleViBe (<a href="http://orcid.org/0009-0008-9399-8581">Vishakh</a>)
**Repository:** https://github.com/DjentleViBe/PIRO.git
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
