### 2025-08-06 #8750 @Brookluo
@DjentleViBe Hi, I am reviewing your package and having some issues with installation:

First, you should specify the required compiler and builder tools in the installation script and check if they are available, installing them if not.

Link to the releases here is void:
https://github.com/DjentleViBe/PIRO/wiki/Getting-Started
Consider changing to:
https://github.com/DjentleViBe/PIRO/releases

You used a pre-compiled zip file for executables, but this can be problematic with different OS environments. When running the `makeall.sh`, it actually creates a tarball (`Linux.tar.gz`) inside the `Executables` directory, but without a real executable `PIRO`.

In your documentation, the `Getting Started` section and `How to build` section seem to contradict each other on `./bin` directory and the contents of it. Are we compiling to make the `PIRO` ourselves, or should we use the `PIRO` provided by your tarball?

Currently, I am not able to install the program using the information on the wiki, so please make sure the installation guide is correct.

I am installing it on Ubuntu 25.04 and am happy to provide the installation messages upon request.
