### 2025-08-07 DjentleViBe/PIRO#34 @Brookluo
[JOSS Review]: failed to install the package with current instructions
I can't install the package on my computer with Ubuntu 25.04 and no GPU. For detailed info, please see my comment here: [#8750](https://github.com/openjournals/joss-reviews/issues/8750#issuecomment-3161894674)
