### 2025-08-07 DjentleViBe/PIRO#34 @DjentleViBe
At the moment, the requirements script forces installation of the latest required dependencies for running. Device list (CPU and GPU) should still be printed (through ```PIRO_devices_LIN```) which is also handled in the [Piro_requirements_run.sh](https://github.com/DjentleViBe/PIRO/wiki/Requirements). Could you please share the error which you get when you follow this step?

The tarball which is created in the ```Executables``` directory uses the binary that is created after running ```./makeall.sh```. During execution of this script the folder downloaded in the ```Getting started``` section is replaced. This is only for building.

For running: Getting started - Requirements(Running) - How-to run
For building Getting started - Requirements(Building) - How-to build
For testing: Getting started - Requirements - Testing

Releases section link is updated.
