### 2025-08-08 DjentleViBe/PIRO#34 @Brookluo
Sharing a bit of context: I am using a fresh Ubuntu 25.04 docker image to test your package. My host machine is an ARM-based Mac. Here is the error message.
```bash
./Piro_requirements_run.sh: line 15: lspci: command not found
Please note down the Device and Platform number of the hardware where the simulation is intended to be run.
rosetta error: failed to open elf at /lib64/ld-linux-x86-64.so.2
./Piro_requirements_run.sh: line 62:  3928 Trace/breakpoint trap   ./PIRO_devices_LIN
```
The first error is caused by missing `lspci` cmd, so I'd recommend including `pciutils` in the installation script, in addition to the specific compiler and building tools that I suggested in my previous comments.
The second error is caused by your executable (or binary) `PIRO_devices_LIN`. Since this binary was precompiled, I think you might want to create a binary per OS. However, that is still sub-optimal. It would be great if we can compile this binary locally (i.e., either through docker multi-stage build that completely gets rid of the host OS issue, or sharing the makefile).

Some further thoughts on this: you claimed the software to be cross-platform, but the supported architectures (i.e., x86, ARM, mips, etc.) weren't specified. I did a test with linux/x86 to install the package, and the second issue didn't show up.

So, please add more information on the supported platform and architecture to run your software.
