### 2025-08-08 DjentleViBe/PIRO#34 @Brookluo
Here is the output by using a linux/x86 platform output:

[install.log](https://github.com/user-attachments/files/21693355/install.log)
