### 2025-08-09 DjentleViBe/PIRO#34 @DjentleViBe
Thank you so much for your detailed feedback.

- ```pciutils``` has been included now in the installation script
- Documentation now specifies supported OS and architectures in requirements.

I believe OpenCL is not supported on ARM hardware and the error is not an OS issue. The binaries provided are compiled on x64 for each OS type.
The make file for PIRO_devices_LIN is under the [devices_list](https://github.com/DjentleViBe/PIRO/tree/device_list) branch.  It didn’t seem necessary to compile this file with every release since it’s only used for initialisation and retrieving device IDs.

I attempted a Docker-based solution, but found that reliable access to OpenCL devices was inconsistent across different OS. I shall update the installation script to explicitly specify the required compiler and build tools.
