### 2025-08-09 DjentleViBe/PIRO#34 @DjentleViBe
> Here is the output by using a linux/x86 platform output:
>
> [install.log](https://github.com/user-attachments/files/21693355/install.log)

Could you please share the hardware details of the machine?
