### 2025-08-09 b0e18c0
make and run device linux
