+++ makedevice.sh
@@ -1,0 +1,5 @@
+make -f Makefile.devices clean
+make -f Makefile.devices
+make -f Makefile.devices setup
+
+chmod +x .$SCRIPT_DIR/PIRO_devices_LIN
