### 2025-08-09 a49beca
make device windows
