+++ rundevice.sh
@@ -1,2 +1,3 @@
 echo "Please note down the Device and Platform number of the hardware where the simulation is intended to be run."
-./scripts/linux/PIRO_devices_LIN
+if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+    ./scripts/linux/PIRO_devices_LIN
