### 2025-08-09 e475de5
run device script
