+++ rundevice.sh
@@ -1,3 +1,4 @@
 echo "Please note down the Device and Platform number of the hardware where the simulation is intended to be run."
 if [[ "$OSTYPE" == "linux-gnu"* ]]; then
     ./scripts/linux/PIRO_devices_LIN
+fi
