### 2025-08-10 4a72611
make and run device linux
