+++ makedevice.sh
@@ -2,4 +2,4 @@ make -f Makefile.devices clean
 make -f Makefile.devices
 make -f Makefile.devices setup
 
-chmod +x .$SCRIPT_DIR/PIRO_devices_LIN
+chmod +x ./scripts/linux/PIRO_devices_LIN
