### 2025-08-11 7e279c5
arch linux nvidia opencl
