### 2025-08-11 a6b00ba
sudo command
