+++ scripts/linux/Piro_requirements_run.sh
@@ -1,9 +1,15 @@
 #!/bin/bash
 SCRIPT_DIR="$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
-tools_ubuntu=("wget" "git=1:2.34.1-1ubuntu1.15" "clinfo=clinfo=3.0.21.02.21-1" "rsync" "pciutils" "ocl-icd-opencl-dev=2.2.14-3" "pocl-opencl-icd=1.8-3")
-tools_suse=("wget" "git-2.50.1-2.1" "which" "clinfo-3.0.25.02.14-1.1" "rsync" "pciutils" "ocl-icd-devel-2.3.3-1.1" "pocl-7.0-1.1" "pocl-devel-7.0-1.1")
+tools_ubuntu=("wget" "git=1:2.34.1-1ubuntu1.15" "clinfo=3.0.21.02.21-1" "rsync" "pciutils"  "ocl-icd-opencl-dev=2.2.14-3" "pocl-opencl-icd=1.8-3")
+tools_suse=("wget" "git-2.50.1-2.1" "which" "rsync" "pciutils" "clinfo-3.0.25.02.14-1.1" "ocl-icd-devel-2.3.3-1.1" "pocl-7.0-1.1" "pocl-devel-7.0-1.1")
 tools_fedora=("wget" "git-2.50.1-1.fc42" "which" "clinfo-3.0.23.01.25-7.fc42" "rsync" "pciutils" "opencl-headers-3.0-32.20241023git4ea6df1.fc42" "pocl-6.0-6.fc42")
 tools_arch=("wget" "git" "clinfo" "rsync" "pciutils" "ocl-icd" "pocl")
+
+if command -v sudo &>/dev/null; then
+    SUDO="sudo"
+else
+    SUDO=""
+fi
 
 # Run as root check
 if [ "$EUID" -ne 0 ]; then
@@ -30,10 +36,10 @@ check_tools() {
         else
             cmd="$tool"             # No version specified
         fi
-        if command_exists "$tool"; then
-            echo "$tool is installed."
+        if command_exists "$cmd"; then
+            echo "$cmd is installed."
         else
-            echo "$tool is NOT installed."
+            echo "$cmd is NOT installed."
             missing_ref+=("$tool")
         fi
     done
@@ -53,10 +59,11 @@ if [ -f /etc/debian_version ]; then
         apt install -y "$t"
     done
 
-    # Optional: NVIDIA OpenCL runtime if NVIDIA GPU detected
     if lspci | grep -i nvidia > /dev/null; then
         echo "NVIDIA GPU detected, installing NVIDIA OpenCL ICD..."
-        # apt install -y libcompute-nvidia-570
+        DRIVER_VERSION=$(nvidia-smi --query-gpu=driver_version --format=csv,noheader | head -n1)
+        DRIVER_MAJOR=$(echo "$DRIVER_VERSION" | cut -d. -f1)
+        apt install -y "libnvidia-compute-${DRIVER_MAJOR}"
     fi
 
 elif [ -f /etc/redhat-release ]; then
@@ -101,7 +108,7 @@ elif [ -f /etc/arch-release ]; then
 
         # Install the package
         echo "Installing $PACKAGE_NAME..."
-        sudo pacman -U --noconfirm "$PACKAGE_NAME" || { echo "Installation failed"; rm -f "$PACKAGE_NAME"; exit 1; }
+        $SUDO pacman -U --noconfirm "$PACKAGE_NAME" || { echo "Installation failed"; rm -f "$PACKAGE_NAME"; exit 1; }
 
         # Delete the package file after install
         echo "Deleting package file $PACKAGE_NAME..."
