### 2025-08-11 c24d2ef
overwrite nvidia instalation for arch linux
