### 2025-08-11 92e8b96
fedora nvidia
