+++ scripts/linux/Piro_requirements_run.sh
@@ -1,8 +1,11 @@
 #!/bin/bash
 SCRIPT_DIR="$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
-tools_ubuntu=("wget" "git=1:2.34.1-1ubuntu1.15" "clinfo=3.0.21.02.21-1" "rsync" "pciutils"  "ocl-icd-opencl-dev=2.2.14-3" "pocl-opencl-icd=1.8-3")
+tools_ubuntu_2204=("wget" "git=1:2.34.1-1ubuntu1.15" "clinfo=3.0.21.02.21-1" "rsync" "pciutils" "ocl-icd-opencl-dev=2.2.14-3" "pocl-opencl-icd=1.8-3")
+tools_ubuntu_generic=("wget" "git" "clinfo" "rsync" "pciutils"  "ocl-icd-opencl-dev" "pocl-opencl-icd")
 tools_suse=("wget" "git-2.50.1-2.1" "which" "rsync" "pciutils" "clinfo-3.0.25.02.14-1.1" "ocl-icd-devel-2.3.3-1.1" "pocl-7.0-1.1" "pocl-devel-7.0-1.1")
-tools_fedora=("wget" "git-2.50.1-1.fc42" "which" "clinfo-3.0.23.01.25-7.fc42" "rsync" "pciutils" "opencl-headers-3.0-32.20241023git4ea6df1.fc42" "pocl-6.0-6.fc42")
+tools_suse_generic=("wget" "git" "which" "rsync" "pciutils" "clinfo" "ocl-icd-devel" "pocl" "pocl-devel")
+tools_fedora_42=("wget" "git-2.50.1-1.fc42" "which" "clinfo-3.0.23.01.25-7.fc42" "rsync" "pciutils" "opencl-headers-3.0-32.20241023git4ea6df1.fc42" "pocl-6.0-6.fc42")
+tools_fedora_generic=("wget" "git" "which" "clinfo" "rsync" "pciutils" "opencl-headers" "pocl")
 tools_arch=("wget" "git" "clinfo" "rsync" "pciutils" "ocl-icd" "pocl")
 
 if command -v sudo &>/dev/null; then
@@ -19,6 +22,7 @@ fi
 
 missing_tools=()
 echo "Checking commands..."
+
 command_exists() {
     command -v "$1" >/dev/null 2>&1
 }
@@ -54,7 +58,12 @@ check_tools() {
 if [ -f /etc/debian_version ]; then
     echo "Debian/Ubuntu detected."
     apt update
-    check_tools tools_ubuntu missing_tools
+    OS_VERSION=$(grep '^VERSION=' /etc/os-release | cut -d'"' -f2)
+    if [[ "$OS_VERSION" == 22.04* ]]; then
+        check_tools tools_ubuntu_2204 missing_tools
+    else
+        check_tools tools_ubuntu_generic missing_tools
+    fi
     for t in "${missing_tools[@]}"; do
         apt install -y "$t"
     done
@@ -69,14 +78,25 @@ if [ -f /etc/debian_version ]; then
 elif [ -f /etc/redhat-release ]; then
     echo "Red Hat/CentOS/Fedora detected."
     yum update -y
-    check_tools tools_fedora missing_tools
+    OS_VERSION=$(grep '^VERSION_ID=' /etc/os-release | cut -d'=' -f2 | tr -d '"')
+    if [[ "$OS_VERSION" == "42" ]]; then
+        check_tools tools_fedora_42 missing_tools
+    else
+        check_tools tools_fedora_generic missing_tools
+    fi
     for t in "${missing_tools[@]}"; do
         yum install -y "$t"
     done
-
+    $SUDO ln -s $(which g++-14) /usr/bin/g++
[…]
