### 2025-08-11 d9248cc
opensuse nvidia
