### 2025-08-11 953ed88
added unsupported linux distribution comment
