### 2025-08-11 d60138c
check architecture type
