### 2025-08-11 f000dca
64 bit check
