### 2025-08-11 2b4f0bb
32 bit check in docker
