### 2025-08-11 DjentleViBe/PIRO#34 @DjentleViBe
> I'm using a Mac, but with cross-platform ubuntu image to emulate linux/x86 under the hood. Is the error related to hardware even with a full emulated docker image?

I think so, since OpenCL calls still require hardware interaction and passthrough. Using a container will still need communication with the host kernel and since OpenCL is not supported on Mac ARM it will fail.

- I have added a [Platforms](https://github.com/DjentleViBe/PIRO/wiki/Platforms) section to track the compatibility of supplied binaries.
- [Builder tools](https://github.com/DjentleViBe/PIRO/wiki/Requirements#building) will now install required dependencies based on OS. I have added a check for x86_64, it explicitly checks the architecture.

For building - 32 bit dependency is left to user's discretion, for running - the script will come to a hard stop.
I have tested the builder script for various x86_64 host/container combinations. If other varieties are desired it can be  populated into the ```./Piro_Requirements_build.sh```.

32 bit support has been dropped by majority vendors. I recommend moving to a 64bit system, I could imagine that especially where vectorisation is important you might loose quite some performance, since you cannot use all AVX/AVX2 registers on x86-32.
