### 2025-08-12 DjentleViBe/PIRO#34 @Brookluo
Sure, but I wouldn't expect it to fail during installation, right? Unless there are some hardware checks to intentionally halt the compilation. If that is the case, I would also expect to see some helpful messages. From another perspective, since recent Macs are all using ARM processors without OpenCL support (they have their own Metal framework) and Apple will discontinue software updates for Intel-based Macs starting next year, have you considered removing the support for Macs? Otherwise, it's extra work to maintain and test your software on another gradually deprecating system. But it's totally up to you.

Anyways, I'll try to install it on an x86 platform.
