### 2025-08-12 DjentleViBe/PIRO#34 @DjentleViBe
> Sure, but I wouldn't expect it to fail during installation, right? Unless there are some hardware checks to intentionally halt the compilation. If that is the case, I would also expect to see some helpful messages. From another perspective, since recent Macs are all using ARM processors without OpenCL support (they have their own Metal framework) and Apple will discontinue software updates for Intel-based Macs starting next year, have you considered removing the support for Macs? Otherwise, it's extra work to maintain and test your software on another gradually deprecating system. But it's totally up to you.
>
> Anyways, I'll try to install it on an x86 platform.

If the binaries are provided for the platform then it should be fine.  I have added these checks in the latest commit.
True. I have considered removing the support for Intel Macs. Tbh the installation for Intel Macs was super straightforward, everything came bundled in with the OS and no tedious instructions were necessary, which is why I had kept it as my main development machine. But yeah, I also don't see a future for that platform.

It would be interesting to make this compatible with the aarch64 architecture as well. However, I don’t have the hardware to test it, so determining the library names for respective OS and performing validation is out of scope for now.
