### 2025-10-04 DjentleViBe/PIRO#36 @Yurlungur
[JOSS-Review] feedback on paper
Hi @DjentleViBe I am a reviewer for your JOSS paper for openjournals/joss-reviews#8750 . I wanted to put some comments here regarding the paper itself. Overall I liked it, especially the detailed description of scaling and how you choose the hash key. However, I do think some things should be addressed (in addition to the code/repo review) before acceptance in to JOSS:

- [ ] It is not clear from the prose what $\lambda$ is supposed to represent. I am also confused by the notation of `:` in the equation defining it.
- [ ] Using $nnz$ to represent the number of non-zero elements is hard to parse, as it's three letters for one variable. I would recommend instead using a single letter, perhaps $e$.
- [ ] The scope of the code does not seem captured by the writeup. For example, I see in the README discussions of operators such as a Lapplacian. How are these discretized? At what order? Your library is clearly a sparse solver designed for elliptic and parabolic PDEs but you only discuss the matrix method.
- [ ] The discussion of the state of the field is rather incomplete:
  - [ ] I was surprised that there was no discussion of the larger GPU-capable sparse linear solvers in the community. For example, [PETSc](https://petsc.org/release/), and [HYPRE](https://computing.llnl.gov/projects/hypre-scalable-linear-solvers-multigrid-methods) are vendor agnostic, GPU capable and specifically designed for sparse linear solves in PDEs, much like your library. More linear algebra focused projects like [Eigen](https://github.com/PX4/eigen), [MAGMA](https://icl.utk.edu/magma/), and [Kokkos-Kernels](https://kokkos.org/kokkos-kernels/docs/) have also earned significant pedigree. Your work is a great supplement/compliment to the more established solvers. But this community should be better described in your paper.
  - [ ] I was also surprised by the claim that using a hash table for sparse matrix representations is novel. This idea has been re-discovered a number of times. For example, it was [implemented in PETSc in 2007](https://link.springer.com/chapter/10.1007/978-3-540-75755-9_107) and discussed in an IEEE paper [as recently as 2022](https://ieeexplore.ieee.org/abstract/document/9851653). Even Python's scipy has an [implementation](https://docs.scipy.org/doc/scipy/reference/generated/scipy.sparse.dok_matrix.html#scipy.sparse.dok_matrix)
- [ ] Related to the above point, the references need to be significantly expanded to include these relevant works.
- [ ] Figure 2 is very difficult to read. Can you make it larger?
- [ ] The timings in Table 1 seem pretty hard for me to interpret. The timings for, e.g., `ROWSEP` seem implausibly slow. Do you have an explanation for this?
- [ ] Also for table 1, $7^3$ is a very small matrix. Why was that size chosen for the timings?
