+++ paper.bib
@@ -11,6 +11,83 @@
   doi     = {10.1109/InPar.2012.6339602}
 }
 
+@misc{hypre,
+  author = {Lawrence Livermore National Laboratory},
+  title = {HYPRE: Scalable Linear Solvers and Multigrid Methods},
+  year = {2007},
+  publisher = {Lawrence Livermore National Laboratory},
+  journal = {},
+  url = {https://computing.llnl.gov/projects/hypre-scalable-linear-solvers-multigrid-methods}
+}
+
+@misc{petsc,
+  author = {UChicago Argonne, LLC},
+  title = {PETSc: Portable, Extensible Toolkit for Scientific Computation},
+  year = {2025},
+  publisher = {UChicago Argonne, LLC},
+  journal = {},
+  url = {https://petsc.org/release/}
+}
+
+@misc{eigen,
+  author = {},
+  title = {EIGEN},
+  year = {2025},
+  publisher = {},
+  journal = {},
+  url = {https://github.com/PX4/eigen}
+}
+@misc{magma,
+  author = {},
+  title = {Innovative Computing Laboratory},
+  year = {2025},
+  publisher = {},
+  journal = {},
+  url = {https://icl.utk.edu/magma/}
+}
+
+@misc{magmasparse,
+  author = {},
+  title = {Innovative Computing Laboratory},
+  year = {2025},
+  publisher = {},
+  journal = {},
+  url = {https://icl.utk.edu/projectsfiles/magma/doxygen/_m_a_g_m_a-sparse.html}
+}
+
+@article{aspn,
+author="Aspn{\"a}s, Mats
+and Signell, Artur
+and Westerholm, Jan",
+editor="K{\aa}gstr{\"o}m, Bo
+and Elmroth, Erik
+and Dongarra, Jack
+and Wa{\'{s}}niewski, Jerzy",
+title="Efficient Assembly of Sparse Matrices Using Hashing",
+booktitle="Applied Parallel Computing. State of the Art in Scientific Computing",
+year="2007",
+publisher="Springer Berlin Heidelberg"
+}
+
+@misc{viennacl,
+  author = {TU Wien},
+  title = {ViennaCL},
+  year = {2016},
+  publisher = {},
+  journal = {},
+  url = {https://viennacl.sourceforge.net}
+}
+
+@misc{du2022opsparsehighlyoptimizedframework,
+      title={OpSparse: a Highly Optimized Framework for Sparse General Matrix Multiplication on GPUs}, 
+      author={Zhaoyang Du and Yijin Guan and Tianchan Guan and Dimin Niu and Linyong Huang and Hongzhong Zheng and Yuan Xie},
+      year={2022},
+      eprint={2206.07244},
+      archivePrefix={arXiv},
+      primaryClass={cs.DC},
+      url={https://arxiv.org/abs/2206.07244}, 
+}
+
 @misc{cusparse,
   author = {NVIDIA},
   title = {cuSPARSE Library},
