+++ paper.md
@@ -25,27 +25,30 @@ bibliography: paper.bib
 Compressed Sparse Row (CSR) is a space efficient way for storing sparse-matrices. It is particularly useful in situations where the matrix has many zero elements. Parallellising CSR matrix operations requires special treatment as not all elements in the matrix are readily available for independent processing. 
 One solution is to process each row separately, making it straightforward to assign different rows to different threads. This approach requires the availability of atomic operations to extract good performance [@ieeepaper2012]. 
 For dynamic graphs stored in CSR, parallel algorithms can insert or delete edges concurrently. Lock-free or fine-grained locking mechanisms are employed to allow multiple threads to update the structure without significant contention[@dyngraphs]. However, this introduces complexity in managing numerous locks, increasing the risk of deadlock or priority inversion if not designed meticulously [@ieeepaper2024].
-Existing libraries often lack GPU support or are tied to proprietary platforms (e.g., CUDA [@cusparse]). PIRO fills this gap by offering a novel solution (HTLF) by representing Sparse Matrices as Hash Tables (HT). It uses a hash function to compute an index (hash key) into an array of slots, where the corresponding value is stored. Insertion, deletion and lookup can be performed in amortized constant time (on average) independent of the number of non zero elements in the table, assuming a good hash function and  well-sized table. It is well known that HT are generally more efficient than search trees or other lookup structures for these operations, especially when fast access is required.
+The landscape of high-performance linear algebra and sparse solvers is diverse. 
+Existing libraries often lack GPU support or are tied to proprietary platforms (e.g., CUDA [@cusparse]). 
+Libraries such as PETSc [@petsc] supports OpenCL through the ViennaCL[@viennacl] backend, however, since ViennaCL primarily targets older OpenCL versions, this approach may lack full compatibility with modern OpenCL features. HYPRE[@hypre] is a library of high performance preconditioners and solvers designed for structured and semi-structured grids but GPU acceleration and dynamic matrix updates are limited. Eigen [@eigen] is a highly mature C++ template library for linear algebra which is CPU-focused. Magma [@magma] is highly optimized for dense linear algebra on CUDA, MAGMA-sparse [@magmasparse] is now in Legacy Support Mode (as of December 2024), with no active development.
 
+While hash-based sparse matrix assembly has been established on CPUs [@aspn], adapting it for GPU execution is still a recent challenge [@du2022opsparsehighlyoptimizedframework]. PIRO addresses this gap with HTLF (Hash table with Load Factor), a complementary approach that scales the hash table using a fitting function. It uses a hash function to compute an index (hash key) into an array of slots, where the corresponding value is stored. Insertion, deletion and lookup can be performed in amortized constant time (on average) independent of the number of non zero elements in the table, assuming a good hash function and well-sized table. It is well known that HT are generally more efficient than search trees or other lookup structures for these operations, especially when fast access is required.
 
 Additionally the software offers:
 
-- __Cross-platform CPU / GPU operation__ via OpenCL.
-- __Modular equation__ solver.
+- __Cross-platform CPU / GPU operation__ for dense and CSR based matrix matrices via OpenCL.
+- __Modular equation__ solver for Partial Differential Equations.
 - __Post processing__ export function for viewing results (e.g Paraview).
 - __Benchmarking__ tools and diagnostics. 
 
 # Algorithm - HTLF
[…]
