### 2025-10-05 DjentleViBe/PIRO#36 @DjentleViBe
@Yurlungur Thank you so much for accepting to review my paper and your comments regarding it. I’m happy to hear you liked the description. I have addressed the following in the latest commit:
- Clarification on what λ represents. Better notation for its corresponding equation.
- Using e stead of nnz.
- Description of the mentioned communities, their current status and limitations for example, and how this code can act as a supplement to existing designs.
- I suppose I wanted to say in the case of GPU (OpenCL) specifically.
- Figure 2 is made bigger.
- The rowsep method was slow because it is repeatedly copying the solution for each row back to the host and then the updated matrix back to device, for each gaussian elimination step.
- The sensitivity analysis was done based on time feasibility of the rowsep method.

Regarding the content of the paper itself, I was not sure of the scope of the content to add. The production-ready research method is limited to the algorithm itself, the rest of the features I felt were mostly research application oriented. In its current state, the PDE solvers use Finite Difference Methods due to the structured nature of the gird. Semi-structured (Octree-based) is still under development (Octree branch) as this also needs to be performed exclusively on ```device``` to prevent back and forth copying. For fully unstructured, the existing open-source solutions -  [SnappyHexMesh](https://www.openfoam.com/documentation/user-guide/4-mesh-generation-and-conversion/4.4-mesh-generation-with-the-snappyhexmesh-utility), [SALOME](https://www.salome-platform.org), [gmsh](https://gmsh.info/doc/texinfo/gmsh.html), [netgen](https://sourceforge.net/projects/netgen-mesher/) are limited to tetrahedral / hex cell types. I would prefer to have a polyhedral type (mesh generation method is in proof of concept stage). The ```main```branch will be extended to incorporate this.
Would it be appropriate to include it as ongoing development efforts in the paper or is it only limited to completed features?
