### 2025-10-05 DjentleViBe/PIRO#37 @Yurlungur
[JOSS-review] instructions for end-to-end example a little confusing
Another review comment for openjournals/joss-reviews#8750. I notice that the [end-to-end example](https://github.com/DjentleViBe/PIRO/wiki/end-to-end-example) has instructions for installation that include *both* cloning the repo *and* downloading a release. Is this intentional? Should it not say do one or the other?
