### 2025-10-05 DjentleViBe/PIRO#36 @Yurlungur
Thanks, @DjentleViBe the paper looks much improved now! I think you addressed al my comments. Regarding the scope of the paper, from my perspective it's fine to include features under development, so long as their state is clear. Perhaps in a brief future work section. But I don't think it's necessary. I'm happy to sign off on the paper portion of the review now. :+1:
