### 2025-10-05 DjentleViBe/PIRO#37 @DjentleViBe
This is intentional. The compressed files provides pre-compiled binaries for running. Platform and details are [here](https://github.com/DjentleViBe/PIRO/wiki/Platforms). Tracking binaries in repo would have been cumbersome. If building, this step is not required.
