### 2025-10-05 DjentleViBe/PIRO#36 @Yurlungur
> The rowsep method was slow because it is repeatedly copying the solution for each row back to the host and then the updated matrix back to device, for each gaussian elimination step.

Is that necessary? Could rowsep be done entirely on device? If so, I would at least caveat this limitation in the timing, or simply exclude the timings and only show the asymptotic costs.
