### 2025-10-05 DjentleViBe/PIRO#37 @Yurlungur
Ah ok. It should be clarified that the binaries can be built or downloaded then.

Also on this page [here](https://github.com/DjentleViBe/PIRO/wiki/how-to-run) it looks like you duplicate the instruction `runall` twice, but one of them should be `makeall`.
