### 2025-10-05 #8750 @Yurlungur
@danielskatz for this bullet point

> Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

The paper does contain timing data for the author's GPU, which I do not have a copy of. I am inclined to just check this box, assuming I believe the timing data makes sense. (We are discussing that in a linked issue.) Does that seem reasonable?
