### 2025-10-05 DjentleViBe/PIRO#38 @Yurlungur
[JOSS-review] Builds fail for AARCH64
Another review for openjournals/joss-reviews#8750. I tried building on my linux VM on my Apple laptop. It looks like you use `immintrin.h`. I believe this is `x86-64` specific, and indeed the header isn't available on my system. Perhaps you could optionally disable these intrinsics or use a portable library?
