### 2025-10-05 DjentleViBe/PIRO#36 @DjentleViBe
In the gaussian elimination step, some zero values become non-zero and non-zeros become zero between iterations. The latest update has to be synchronised for the next step. The only way to perform entirely on device is to copy the whole matrix which would be the same as dense format.
