### 2025-10-05 DjentleViBe/PIRO#37 @Yurlungur
Thanks for the explanation. I think significantly more explanation is needed in each section of the wiki explaining what your commands and scripts do. I am used to working with gnu autotools or cmake and your home-brewed build system is pretty opaque to me.
