### 2025-10-05 DjentleViBe/PIRO#34 @Yurlungur
Hi checking in. I'm also having trouble. A few comments/questions:

1. Is OpenCL not supported for arm-based *linux*? I just want to build and run with CPU-only shared memory in a VM. The VM does not know that it's on apple, it just knows the microcode is AARCH64. Surely that should be possible, unless all arm builds are excluded?
2. I tried building on an x86_64 machine I have access to (but not root) and I am still having trouble. I downloaded the Linux release and tried to run the advection example. Here is the error message I'm getting from the log:
```
(base) jonahm@cn133:$ cat logs/AdvEq.ini.txt
./bin/PIRO: /lib64/libm.so.6: version `GLIBC_2.29' not found (required by ./bin/PIRO)
./bin/PIRO: /lib64/libc.so.6: version `GLIBC_2.38' not found (required by ./bin/PIRO)
./bin/PIRO: /lib64/libc.so.6: version `GLIBC_2.34' not found (required by ./bin/PIRO)
```
it would be helpful to know *exactly* what build environment you're expecting.
3. Your requirements scripts require root. Not all systems that people will run on support root. For example, the x86_64 machine I mention above is a supercomputer that I am a resident on. I do not have root access there. Is a home-directory build possible?

The good news is I *can* successfully build and run your tests. Though I don't know why the tests can be built and run and the code cannot.
