### 2025-10-05 DjentleViBe/PIRO#39 @Yurlungur
[JOSS review]
Another question for openjournals/joss-reviews#8750. I looked in the CI and there do not appear to be any automated tests that are passing. Did I miss something?
