### 2025-10-05 DjentleViBe/PIRO#40 @Yurlungur
[JOSS-review] How is an equation to solve specified?
Another question for openjournals/joss-reviews#8750. I do not understand the discussion in configuration. How do I specify the equation I want to solve? Also are nonlinear PDEs possible? Or only linear?
