### 2025-10-05 DjentleViBe/PIRO#34 @Yurlungur
> Could you share details about your Linux distribution? I believe it's a missing library from an older distribution.

It's redhat, but it's on a supercomputer, so the environment is managed through modules. If there's a specific version of glibc it may well be available if I know the versions.

> It should work without root on Windows, I have removed the privilege condition. Could you please check if it works?

Just to check, also on Linux?
