### 2025-10-06 DjentleViBe/PIRO#34 @Yurlungur
> 1. If you are using Apple Silicon I don't think OpenCL is not [supported](https://support.apple.com/en-us/101525). This is required even for CPU-only tasks.

I think what you're saying is you only support x86, which is fine. But that's different from apple silicon. Apple silicon extends and implements aarch64. So it should be possible to build opencl for a generic arm system. ARM is much more than just apple.
