### 2025-10-06 69a6949
ubuntu + arch rootless install
