+++ scripts/linux/Piro_requirements_run_rootless.sh
@@ -1,0 +1,457 @@
+#!/bin/bash
+SCRIPT_DIR="$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
+tools_ubuntu_generic=("wget" "git" "clinfo" "rsync" "pciutils" "libpci3" "libkmod2" "pocl" "ocl-icd-opencl-dev" "pocl-opencl-icd" "ocl-icd-libopencl1")
+tools_arch=("wget" "git" "clinfo" "rsync" "pciutils" "ocl-icd" "pocl")
+#tools_suse_generic=("wget" "git" "which" "rsync" "pciutils" "clinfo" "ocl-icd-devel" "pocl" "pocl-devel" "libOpenCL1")
+tools_suse_generic=("pciutils" "clinfo" "ocl-icd-devel" "pocl" "pocl-devel" "libOpenCL1")
+tools_fedora_generic=("wget" "git" "which" "clinfo" "rsync" "pciutils" "opencl-headers" "pocl")
+
+architecture=$(uname -m)
+bits=$(getconf LONG_BIT)
+LOCAL_PREFIX="$(pwd)/local"
+mkdir -p "$LOCAL_PREFIX"
+
+if [[ "$architecture" == "x86_64" && "$bits" == "64" ]]; then
+    echo "System is 64-bit."
+else
+    echo "System is NOT x86_64. (detected architecture: $architecture, bits: $bits)"
+    exit 1
+fi
+
+seen=()
+
+install_ubuntu_package() {
+    local pkg="$1"
+
+    if [[ " ${seen[@]} " =~ " $pkg " ]]; then
+        return
+    fi
+    seen+=("$pkg")
+
+    echo "Resolving dependencies for $pkg ..."
+
+    # Create local directories
+    mkdir -p "$LOCAL_PREFIX"
+
+    # Get all dependencies (including recommended)
+    local all_deps
+    all_deps=$(apt-get --just-print install "$pkg" 2>/dev/null \
+        | awk '/Inst /{print $2}' | sort -u)
+
+    if [[ -z "$all_deps" ]]; then
+        echo "No packages found for $pkg (maybe already downloaded or installed)."
+        return
+    fi
+
+    echo "Downloading packages: $all_deps"
+
+    # Download all .deb files without installing (non-root)
+    apt download $all_deps >/dev/null 2>&1 || {
+        echo "Failed to download $pkg or some dependencies"
+        return
+    }
+
+    # Move downloaded .deb files to LOCAL_PREFIX
+    mv -f ./*.deb "$LOCAL_PREFIX"/ 2>/dev/null || true
+
+    echo "Extracting into $LOCAL_PREFIX ..."
+    find "$LOCAL_PREFIX" -type f -name "*.deb" | while read -r debfile; do
+        dpkg-deb -x "$debfile" "$LOCAL_PREFIX"
+    done
+
+    # Optional cleanup
+    rm -f "$LOCAL_PREFIX"/*.deb
+
+    # Fix OpenCL ICD files if relevant
+    mkdir -p "$LOCAL_PREFIX/etc/OpenCL/vendors"
+    find "$LOCAL_PREFIX" -name "*.icd" -exec cp {} "$LOCAL_PREFIX/etc/OpenCL/vendors/" \;
+
+    echo "Done installing $pkg locally."
+}
+
+install_arch_package() {
+    local root_pkg="$1"
+    local -A seen=()
+    local -a queue=("$root_pkg")
+
+    echo "Building dependency graph..."
+
+    while ((${#queue[@]})); do
+        local pkg="${queue[0]}"
+        queue=("${queue[@]:1}")
+
[…]
