### 2025-10-06 DjentleViBe/PIRO#34 @DjentleViBe
> It's redhat, but it's on a supercomputer, so the environment is managed through modules. If there's a specific version of glibc it may well be available if I know the versions.

The binary was created with glibc 2.42

> Just to check, also on Linux?

I have added this [script](https://github.com/DjentleViBe/PIRO/blob/main/scripts/linux/Piro_requirements_run_rootless.sh) for Linux. It is a bit cumbersome because it's not going to use the OS package managers to draw dependency lists. It should work for Ubuntu, ArchLinux and OpenSuse, hopefully. A follow up question - Do you think the rootless version should be the default? And also, if the packages are installed locally, would it help to have an uninstall script?
