### 2025-10-06 DjentleViBe/PIRO#39 @DjentleViBe
Ahh, Since GitHub Actions required self hosted GPUs, I'm doing it locally with ```make_test.sh``` and hardware flags. Since the installation script is now automated, I could set up a quick workflow.
