### 2025-10-06 DjentleViBe/PIRO#34 @Yurlungur
> > It's redhat, but it's on a supercomputer, so the environment is managed through modules. If there's a specific version of glibc it may well be available if I know the versions.
>
> The binary was created with glibc 2.42

Ah ok. Sounds like I may need to just build from on the machine. I'll try to do so.

> > Just to check, also on Linux?
>
> I have added this [script](https://github.com/DjentleViBe/PIRO/blob/main/scripts/linux/Piro_requirements_run_rootless.sh) for Linux.

Cool thanks for adding that!

> A follow up question - Do you think the rootless version should be the default? And also, if the packages are installed locally, would it help to have an uninstall script?

I don't know... it's probably fine to have both versions. People installing on their laptops will have root, but on HPC systems (which is how people doing bigger calculations are likely to work) they won't.

It might also be useful to just have a list of all the required dependencies for each combination of supported hardware, in case folks want to just install it themselves. I think a likely situation (on HPC for example) is that the relevant GPU drivers are available, but not necessarily the version of OpenCL you need, or the compiler. That kind of thing.
