### 2025-10-06 #8750 @matthewfeickert
@Yurlungur thank you for opening issues up, and linking them back here, and your questions &mdash; I've been in (delayed) international transit for work for the last 24+ hours and so am still catching up on GitHub notifications.

> The paper does contain timing data for the author's GPU, which I do not have a copy of. I am inclined to just check this box, assuming I believe the timing data makes sense. (We are discussing that in a linked issue.) Does that seem reasonable?

@Yurlungur when you say you "do not have a copy", do you mean that you do not have the exact same GPU architecture as is reported in the paper (a AMD Radeon Pro 5300M in this case) or that you do not have access to a GPU at all? If it is just that you do not have access to the exact same GPU then I would not be concerned with reproducing the exact timings but validating the trend of the timing ratios (where the nominal time is set to 1).

@DjentleViBe Do you provide code to reproduce the table for the reviewers? Also please note that algorithm descriptions should be covered in the documentation and not the JOSS paper itself unless it is critical to the motivation of the software (c.f. the [Submitting a paper to JOSS](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) section of the JOSS docuementation). Please also to make sure to properly reference all figures and tables in your paper by their reference ID, and not by relative position terms like "below".
