### 2025-10-06 #8750 @matthewfeickert
👋 @Brookluo can you please update us on how the review is going, as you haven't commented on Issue https://github.com/DjentleViBe/PIRO/issues/34 for some time. I see from it that there are still instillation issues, but as you have multiple other outstanding evaluation points on your checklist (https://github.com/openjournals/joss-reviews/issues/8750#issuecomment-3161732615) it would be good to know if there are other things that you're able to evaluate (though I appreciate that if you can't install the software the rest of the review may be effectively blocked).
