### 2025-10-06 DjentleViBe/PIRO#34 @matthewfeickert
I will note that while not required, in the past we have found that providing [Pixi](https://pixi.sh/) workspaces with install `tasks` for CUDA enabled software has made reproduction for reviewers considerably easier. Pixi also operates fully in user space.
