+++ .github/workflows/test-build.yaml
@@ -1,0 +1,25 @@
+name: Building
+
+on:
+  push:
+    branches: [ TESTING ]
+  pull_request:
+    branches: [ TESTING ]
+
+jobs:
+  build-test:
+    runs-on: ${{ matrix.os }}
+
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest]
+        arch: [x64]
+      steps:
+        - name: Checkout
+          uses: actions/checkout@v4
+        
+        - name: Install dependencies (Linux)
+          if: runner.os == 'Linux'
+          run: bash scripts/linux/Piro_requirements_run_rootless.sh
+                bash ./maketest.sh
