+++ .github/workflows/test-build.yaml
@@ -15,11 +15,11 @@ jobs:
       matrix:
         os: [ubuntu-latest]
         arch: [x64]
-      steps:
-        - name: Checkout
-          uses: actions/checkout@v4
-        
-        - name: Install dependencies (Linux)
-          if: runner.os == 'Linux'
-          run: bash scripts/linux/Piro_requirements_run_rootless.sh
-                bash ./maketest.sh
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+      
+      - name: Install dependencies (Linux)
+        if: runner.os == 'Linux'
+        run: bash scripts/linux/Piro_requirements_run_rootless.sh
+              bash ./maketest.sh
