+++ .github/workflows/test-build.yaml
@@ -22,5 +22,7 @@ jobs:
       - name: Install dependencies (Linux)
         if: runner.os == 'Linux'
         run: |
+          sudo apt-get update
+          sudo apt-get install -y ocl-icd-opencl-dev
           bash scripts/linux/Piro_requirements_run_rootless.sh
           bash ./maketest.sh -d 0 -p 0
