### 2025-10-07 aba0272
run devices
