### 2025-10-07 d7b6b10
paper pdf only on main branch
