+++ .github/workflows/draft-pdf.yaml
@@ -1,5 +1,7 @@
 name: Draft PDF
-on: [push]
+on:
+  push:
+    branches: [ main ]
 
 jobs:
   paper:
