### 2025-10-07 5a68f0e
gpucheck for ubuntu removed
