### 2025-10-07 4e9bac6
pointer correction cpu
