+++ .github/workflows/test-build.yaml
@@ -1,4 +1,4 @@
-name: Building
+name: Build and Test
 
 on:
   push:
@@ -7,22 +7,44 @@ on:
     branches: [ TESTING ]
 
 jobs:
-  build-test:
+  rootful-setup:
     runs-on: ${{ matrix.os }}
-
     strategy:
-      fail-fast: false
       matrix:
         os: [ubuntu-latest]
-        arch: [x64]
+        arch: [x64]  # add architectures here
     steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-      
-      - name: Install dependencies (Linux)
-        if: runner.os == 'Linux'
+      - uses: actions/checkout@v4
+
+      - name: Install dependencies (rootful)
         run: |
+          echo "Running on architecture: ${{ matrix.arch }}"
           sudo apt-get update
           sudo scripts/linux/Piro_requirements_run_rootful.sh
-          ./scripts/linux/PIRO_devices_LIN
-          bash ./maketest.sh -d 0 -p 0
+
+      - name: Setup devices
+        run: ./scripts/linux/PIRO_devices_LIN
+
+      - name: Build and test
+        run: bash ./maketest.sh -d 0 -p 0
+
+  rootless-setup:
+    runs-on: ${{ matrix.os }}
+    needs: rootful-setup
+    strategy:
+      matrix:
+        os: [ubuntu-latest]
+        arch: [x64]  # match the matrix
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Install dependencies (rootless)
+        run: |
+          echo "Running rootless on architecture: ${{ matrix.arch }}"
+          bash scripts/linux/Piro_requirements_run_rootless.sh
+
+      - name: Setup devices
+        run: ./scripts/linux/PIRO_devices_LIN
+
+      - name: Build and test
+        run: bash ./maketest.sh -d 0 -p 0
