### 2025-10-07 #8750 @Yurlungur
>  If it is just that you do not have access to the exact same GPU then I would not be concerned with reproducing the exact timings but validating the trend of the timing ratios (where the nominal time is set to 1).

Yep that's what I meant. I do have access to a GPU system and am working with @DjentleViBe on getting things built/running on that system. Sounds good, thanks. :+1:
