+++ scripts/linux/Piro_requirements_run_rootless.sh
@@ -2,8 +2,7 @@
 SCRIPT_DIR="$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
 tools_ubuntu_generic=("wget" "git" "clinfo" "rsync" "pciutils" "libpci3" "libkmod2" "pocl" "ocl-icd-opencl-dev" "pocl-opencl-icd" "ocl-icd-libopencl1")
 tools_arch=("wget" "git" "clinfo" "rsync" "pciutils" "ocl-icd" "pocl")
-#tools_suse_generic=("wget" "git" "which" "rsync" "pciutils" "clinfo" "ocl-icd-devel" "pocl" "pocl-devel" "libOpenCL1")
-tools_suse_generic=("pciutils" "clinfo" "ocl-icd-devel" "pocl" "pocl-devel" "libOpenCL1")
+tools_suse_generic=("wget" "git" "which" "rsync" "pciutils" "clinfo" "ocl-icd-devel" "pocl" "pocl-devel" "libOpenCL1")
 tools_fedora_generic=("wget" "git" "which" "clinfo" "rsync" "pciutils" "opencl-headers" "pocl")
 
 architecture=$(uname -m)
