### 2025-10-08 05b414e
local folder inside scripts
