### 2025-10-08 1dad45e
ubuntu cpu
