+++ scripts/linux/Piro_requirements_run_rootless.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 SCRIPT_DIR="$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
-tools_ubuntu_generic=("wget" "git" "clinfo" "rsync" "pciutils" "libpci3" "libkmod2" "pocl" "ocl-icd-opencl-dev" "pocl-opencl-icd" "ocl-icd-libopencl1")
+#tools_ubuntu_generic=("wget" "git" "clinfo" "rsync" "make" "g++" "build-essential" "libc6" "libc6-dev" "pciutils" "libpci3" "libkmod2" "ocl-icd-opencl-dev" "pocl-opencl-icd" "ocl-icd-libopencl1")
+tools_ubuntu_generic=("wget" "git" "clinfo" "rsync" "make" "g++" "build-essential" "libc6" "libc6-dev" "pciutils" "libpci3" "libkmod2" "ocl-icd-opencl-dev" "ocl-icd-libopencl1")
 tools_arch=("wget" "git" "clinfo" "rsync" "pciutils" "ocl-icd" "pocl")
 tools_suse_generic=("wget" "git" "which" "rsync" "pciutils" "clinfo" "ocl-icd-devel" "pocl" "pocl-devel" "libOpenCL1")
 tools_fedora_generic=("wget" "git" "which" "clinfo" "rsync" "pciutils" "opencl-headers" "pocl")
@@ -63,7 +64,7 @@ install_ubuntu_package() {
 
     # Fix OpenCL ICD files if relevant
     mkdir -p "$LOCAL_PREFIX/etc/OpenCL/vendors"
-    find "$LOCAL_PREFIX" -name "*.icd" -exec cp {} "$LOCAL_PREFIX/etc/OpenCL/vendors/" \;
+    # find "$LOCAL_PREFIX" -name "*.icd" -exec cp {} "$LOCAL_PREFIX/etc/OpenCL/vendors/" \;
 
     echo "Done installing $pkg locally."
 }
@@ -193,79 +194,52 @@ if [ -f /etc/debian_version ]; then
         # echo "Downloading and installing $t ..."
         install_ubuntu_package "$t"
     done
+    mkdir -p "$LOCAL_PREFIX/lib/x86_64-linux-gnu"
+    cp -r $LOCAL_PREFIX/usr/lib/ $LOCAL_PREFIX/
+    mkdir -p "$LOCAL_PREFIX/lib64/"
+    cp -r $LOCAL_PREFIX/usr/lib64/ $LOCAL_PREFIX/
+    mkdir -p $LOCAL_PREFIX/usr/lib/x86_64-linux-gnu
+    cp /usr/lib/x86_64-linux-gnu/libgcc_s.so.1 $LOCAL_PREFIX/usr/lib/x86_64-linux-gnu/
+    #mkdir -p $LOCAL_PREFIX/share/pocl/include
+    #cp $LOCAL_PREFIX/usr/share/pocl/include/* $LOCAL_PREFIX/share/pocl/include/.
+    #cp $LOCAL_PREFIX/usr/share/pocl/* $LOCAL_PREFIX/share/pocl/.
+
     echo "Setting up environment variables in ~/.bashrc ..."
+    BASHRC="$HOME/.bashrc"
 
-    grep -qxF "export LOCAL_PREFIX=\"$LOCAL_PREFIX\"" ~/.bashrc || \
-    echo "export LOCAL_PREFIX=\"$LOCAL_PREFIX\"" >> ~/.bashrc
+    # Marker to avoid duplicates
+    MARKER="# --- Local build environment setup ---"
[…]
