+++ .github/workflows/test-build.yaml
@@ -18,7 +18,7 @@ jobs:
       - name: Configure LOCAL_PREFIX
         run: |
           # Use current working directory to avoid permission issues
-          export LOCAL_PREFIX="$(pwd)/scripts/local"
+          export LOCAL_PREFIX="$(pwd)/local"
           mkdir -p "$LOCAL_PREFIX"
           echo "LOCAL_PREFIX=$LOCAL_PREFIX" >> $GITHUB_ENV
           echo "PATH=$LOCAL_PREFIX/usr/bin:$PATH" >> $GITHUB_ENV
