### 2025-10-08 e49af93
clinfo check
