+++ .github/workflows/test-build.yaml
@@ -34,8 +34,9 @@ jobs:
           cat ~/.bashrc
           
       - name: Setup devices
-        run: ./scripts/linux/PIRO_devices_LIN
-            clinfo
+        run: |
+          ./scripts/linux/PIRO_devices_LIN
+          clinfo
 
       - name: Build and test
         run: bash ./maketest.sh -d 0 -p 0
