+++ .github/workflows/test-build.yaml
@@ -36,6 +36,7 @@ jobs:
         run: |
           ./scripts/linux/PIRO_devices_LIN
           clinfo
+          find /home/runner/work/PIRO/PIRO/local/usr/include -name wordsize.h
 
       - name: Build and test
         run: bash ./maketest.sh -d 0 -p 0
