### 2025-10-08 a668eec
verify bits
