+++ .github/workflows/test-build.yaml
@@ -42,6 +42,7 @@ jobs:
       - name: Verify bits
         run: |
           find /home/runner/work/PIRO/PIRO/local/usr/include -name wordsize.h
+          find $LOCAL_PREFIX -type d
           find $LOCAL_PREFIX/usr/include/x86_64-linux-gnu -type d
           cat ~/.bashrc
       - name: Build and test
