### 2025-10-08 7dbcbce
wordsize.h
