### 2025-10-08 788ef24
check command in sysroot
