### 2025-10-09 e573366
force reinstall into local
