### 2025-10-09 197e1a3
run device test
