### 2025-10-09 5b0a826
launching in bash
