+++ .github/workflows/test-build.yaml
@@ -77,7 +77,7 @@ jobs:
           echo Running dependency setup for Windows...
           call scripts\windows\Piro_requirements_run.bat
       - name: Run EXE in cmd
-        shell: cmd
+        shell: bash
         run: |
           scripts\windows\PIRO_devices_WIN.exe
       - name: Build and test (Windows)
