+++ .github/workflows/test-build.yaml
@@ -76,10 +76,16 @@ jobs:
         run: |
           echo Running dependency setup for Windows...
           call scripts\windows\Piro_requirements_run.bat
-      - name: Run EXE in cmd
-        shell: bash
+      - name: Install Intel OpenCL CPU runtime
+        shell: pwsh
         run: |
-          ./scripts/windows/PIRO_devices_WIN.exe
+            choco install intel-opencl-icd -y
+            clinfo
+      - name: Run EXE in cmd
+        shell: cmd
+        run: |
+          echo Launching PIRO_devices_WIN.exe...
+          scripts\windows\PIRO_devices_WIN.exe
       - name: Build and test (Windows)
         shell: bash
         run: |
