### 2025-10-09 37ea8d1
install cpu drivers windows
