+++ .github/workflows/test-build.yaml
@@ -79,7 +79,8 @@ jobs:
       - name: Install Intel OpenCL CPU runtime
         shell: pwsh
         run: |
-            choco install intel-opencl-icd -y
+            Invoke-WebRequest -Uri "https://registrationcenter-download.intel.com/akdlm/IRC_NAS/a8589e7b-70f8-4ef2-bdc3-7306dfb93e92/w_opencl_runtime_p_2025.2.0.768.exe" -OutFile "C:\temp\w_opencl_runtime_p_2025.2.0.768.exe"
+            Start-Process -FilePath "C:\temp\w_opencl_runtime_p_2025.2.0.768.exe" -ArgumentList "/s" -Wait
             clinfo
       - name: Run EXE in cmd
         shell: cmd
