### 2025-10-09 3b027c1
extract msi
