+++ .github/workflows/test-build.yaml
@@ -80,7 +80,12 @@ jobs:
         shell: pwsh
         run: |
             Invoke-WebRequest -Uri "https://registrationcenter-download.intel.com/akdlm/IRC_NAS/a8589e7b-70f8-4ef2-bdc3-7306dfb93e92/w_opencl_runtime_p_2025.2.0.768.exe" -OutFile "C:\temp\w_opencl_runtime_p_2025.2.0.768.exe"
-            Start-Process -FilePath "C:\temp\w_opencl_runtime_p_2025.2.0.768.exe" -ArgumentList "/s" -Wait
+            Start-Process -FilePath "C:\temp\intel_opencl.exe" -ArgumentList "-extract", "C:\temp\intel_opencl" -Wait
+            Get-ChildItem "C:\temp\intel_opencl"
+            $msi = Get-ChildItem "C:\temp\intel_opencl" -Filter "*.msi" | Select-Object -First 1
+            Write-Host "Installing $($msi.FullName)"
+            Start-Process "msiexec.exe" -ArgumentList "/i", "`"$($msi.FullName)`"", "/qn", "/norestart" -Wait
+            choco install clinfo -y
             clinfo
       - name: Run EXE in cmd
         shell: cmd
