### 2025-10-09 fc28ca1
install windows drivers first
