+++ .github/workflows/test-build.yaml
@@ -71,22 +71,21 @@ jobs:
         arch: [x64]
     steps:
       - uses: actions/checkout@v4
-      - name: Install dependencies (Windows)
-        shell: cmd
-        run: |
-          echo Running dependency setup for Windows...
-          call scripts\windows\Piro_requirements_run.bat
       - name: Install Intel OpenCL CPU runtime
         shell: pwsh
         run: |
-            choco install clinfo -y
             Invoke-WebRequest -Uri "https://registrationcenter-download.intel.com/akdlm/IRC_NAS/a8589e7b-70f8-4ef2-bdc3-7306dfb93e92/w_opencl_runtime_p_2025.2.0.768.exe" -OutFile "C:\temp\intel_opencl.exe"
             Start-Process -FilePath "C:\temp\intel_opencl.exe" -ArgumentList "-extract", "C:\temp\intel_opencl" -Wait
             Get-ChildItem "C:\temp\intel_opencl"
             $msi = Get-ChildItem "C:\temp\intel_opencl" -Filter "*.msi" | Select-Object -First 1
             Write-Host "Installing $($msi.FullName)"
             Start-Process "msiexec.exe" -ArgumentList "/i", "`"$($msi.FullName)`"", "/qn", "/norestart" -Wait
-            clinfo
+      
+      - name: Install dependencies (Windows)
+        shell: cmd
+        run: |
+          echo Running dependency setup for Windows...
+          call scripts\windows\Piro_requirements_run.bat
       - name: Run EXE in cmd
         shell: cmd
         run: |
