### 2025-10-09 fdcf50e
error checks
