+++ .github/workflows/test-build.yaml
@@ -74,12 +74,36 @@ jobs:
       - name: Install Intel OpenCL CPU runtime
         shell: pwsh
         run: |
-            Invoke-WebRequest -Uri "https://registrationcenter-download.intel.com/akdlm/IRC_NAS/a8589e7b-70f8-4ef2-bdc3-7306dfb93e92/w_opencl_runtime_p_2025.2.0.768.exe" -OutFile "C:\temp\intel_opencl.exe"
-            Start-Process -FilePath "C:\temp\intel_opencl.exe" -ArgumentList "-extract", "C:\temp\intel_opencl" -Wait
-            Get-ChildItem "C:\temp\intel_opencl"
-            $msi = Get-ChildItem "C:\temp\intel_opencl" -Filter "*.msi" | Select-Object -First 1
-            Write-Host "Installing $($msi.FullName)"
-            Start-Process "msiexec.exe" -ArgumentList "/i", "`"$($msi.FullName)`"", "/qn", "/norestart" -Wait
+          $ErrorActionPreference = "Stop"
+          New-Item -ItemType Directory -Force -Path "C:\temp" | Out-Null
+
+          Write-Host "Downloading Intel OpenCL runtime..."
+          Invoke-WebRequest -Uri "https://registrationcenter-download.intel.com/akdlm/IRC_NAS/a8589e7b-70f8-4ef2-bdc3-7306dfb93e92/w_opencl_runtime_p_2025.2.0.768.exe" -OutFile "C:\temp\intel_opencl.exe"
+
+          Write-Host "Extracting installer (this can take 2–5 minutes)..."
+          Start-Process -FilePath "C:\temp\intel_opencl.exe" -ArgumentList "-extract", "C:\temp\intel_opencl" -Wait -NoNewWindow
+
+          if (-not (Test-Path "C:\temp\intel_opencl")) {
+            Write-Error "Extraction failed: directory not found"
+            exit 1
+          }
+
+          $msi = Get-ChildItem "C:\temp\intel_opencl" -Filter "*.msi" | Select-Object -First 1
+          if (-not $msi) {
+            Write-Error "No MSI found after extraction."
+            Get-ChildItem -Recurse "C:\temp\intel_opencl"
+            exit 1
+          }
+
+          Write-Host "Installing $($msi.FullName)..."
+          $process = Start-Process -FilePath "msiexec.exe" -ArgumentList "/i", "`"$($msi.FullName)`"", "/qn", "/norestart" -PassThru
+          if (-not $process.WaitForExit(300000)) { # 5 minutes
+            Write-Error "Installation timeout."
+            $process.Kill()
+            exit 1
+          }
+
+          Write-Host "Intel OpenCL CPU runtime installed successfully."
       
       - name: Install dependencies (Windows)
         shell: cmd
