+++ .github/workflows/test-build.yaml
@@ -84,6 +84,7 @@ jobs:
           & "C:\Program Files\7-Zip\7z.exe" x "C:\temp\intel_opencl.exe" -o"C:\temp\"
           echo Installing...
           Start-Process "msiexec.exe" -ArgumentList "/i C:\Temp\w_opencl_runtime_p_2025.2.0.768.msi /qn /norestart" -Wait ALLUSERS=1
+          echo Intel OpenCL runtime installed
       - name: Install dependencies (Windows)
         shell: cmd
         run: |
