+++ .github/workflows/test-build.yaml
@@ -84,11 +84,6 @@ jobs:
         run: |
           echo Running dependency setup for Windows...
           call scripts\windows\Piro_requirements_run.bat
-      - name: Run EXE in cmd
-        shell: cmd
-        run: |
-          echo Launching PIRO_devices_WIN.exe...
-          scripts\windows\PIRO_devices_WIN.exe
       - name: Build and test (Windows)
         shell: bash
         run: |
