+++ runall.sh
@@ -77,11 +77,14 @@ if [[ "$MSYSTEM" == "MINGW64" ]] && [[ -d "/c/cygwin64/bin" ]]; then
     if ! grep -q "cygwin64" ~/.bashrc 2>/dev/null; then
         echo 'export PATH="/c/cygwin64/bin:$PATH"' >> ~/.bashrc
         echo "Added Cygwin to ~/.bashrc"
+        NEWPATH="/c/cygwin64/bin"
+        echo 'export PATH="$NEWPATH:$PATH"' >> ~/.bashrc
         source ~/.bashrc
     fi
 fi
-NEWPATH="/c/cygwin64/bin"
-export PATH="$NEWPATH:$PATH"
+
+source ./scripts/linux/env_setup.sh
+# Which OpenCL library is being used
 cp -r ./dependencies/assets ./bin/.
 mkdir -p logs
 
