### 2025-10-13 a78a449
note for ROWSEP
