### 2025-10-15 DjentleViBe/PIRO#34 @Brookluo
@Yurlungur did you successfully install the package on your supercomputer? I think most of the supercomputing clusters would have similar OS / architecture (I'm assuming you are using a CPU cluster with GPGPU accelerator).

> It might also be useful to just have a list of all the required dependencies for each combination of supported hardware, in case folks want to just install it themselves. I think a likely situation (on HPC for example) is that the relevant GPU drivers are available, but not necessarily the version of OpenCL you need, or the compiler. That kind of thing.

I second that idea. I think it would be nice to have a list of required packages for users to check with. What's even better is to make a jupyter notebook to walk through the installation process and running some examples, in addition to your wiki. Users will be more likely to follow the initial setup with an interactive notebook. You can find an example here: http://github.com/einsteintoolkit/jupyter-et/blob/master/tutorial-server/notebooks/CactusTutorial.ipynb
