### 2025-10-16 DjentleViBe/PIRO#34 @Yurlungur
> [@Yurlungur](https://github.com/Yurlungur) did you successfully install the package on your supercomputer? I think most of the supercomputing clusters would have similar OS / architecture (I'm assuming you are using a CPU cluster with GPGPU accelerator).

I haven't yet had a chance to try again. I should soon. That's right---I was working on a heterogeneous testing cluster we have access to for prototyping and testing code on a variety of architectures. I was trying on a machine with a Skylake CPU and a V100 GPU. That's old hardware at this point, but I think the broad concerns will be the same across most systems. (And for example, fortunately, this system is *not* Cray, so doesn't have their opaque build system obfuscating things.)
