### 2025-11-07 DjentleViBe/PIRO#34 @matthewfeickert
> > It might also be useful to just have a list of all the required dependencies for each combination of supported hardware, in case folks want to just install it themselves. I think a likely situation (on HPC for example) is that the relevant GPU drivers are available, but not necessarily the version of OpenCL you need, or the compiler. That kind of thing.
>
> I second that idea. I think it would be nice to have a list of required packages for users to check with.

I would strongly recommend this. In addition to being important to clearly [document the dependencies and requirements](https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions), tools can be viably installed many different ways and do not need to come from a system level package manager. Letting users decided how they are going to reach this environment is important (Example: We have found at JOSS in the past that if CUDA dependencies are required having reviewers get them from conda-forge through either Pixi or conda/mamba significantly simplifies the review process). This also means that there should be no hardcoded assumptions for building of software of where requirements are coming from.
