### 2025-11-07 #8750 @matthewfeickert
👋 @Brookluo @Yurlungur @DjentleViBe, I'm just checking in to on the state of the review.

@Brookluo @Yurlungur, You've both opened up Issues on https://github.com/DjentleViBe/PIRO (👍) and I see that there's been some recent and ongoing discussions between @Yurlungur and @DjentleViBe on Issues and some Issue resolution also, which is great.

It seems that there has been detailed discussion on installation difficulties in https://github.com/DjentleViBe/PIRO/issues/34, so I would suggest that be a primary focus of friction reduction as the review can't progress if the software can't be installed.

> I have been busy lately after my initial attempt to install the package on my VM. I'll give it a try again on a different machine and report back on it. I expect to finish the review by mid-November.

@Brookluo that's fine. 👍 I'll set reminders for all of us for 2 weeks.
