### 2025-11-13 DjentleViBe/PIRO#34 @DjentleViBe
>  There are currently no build instructions or links on https://github.com/DjentleViBe/PIRO/wiki/end-to-end-example, https://github.com/DjentleViBe/PIRO/wiki/Requirements#building, or https://github.com/DjentleViBe/PIRO/wiki/Platforms#building. It exists only on https://github.com/DjentleViBe/PIRO/wiki/how-to-build and is not linked to elsewhere. This information should also not be spread out across multiple pages on an (collectively) unversioned wiki.

I believe https://github.com/DjentleViBe/PIRO/wiki/Requirements had the link to [build](https://github.com/DjentleViBe/PIRO/wiki/how-to-build) at the very top of the page. I have renamed the Building section to ```Build requirements``` to avoid confusion and added [build](https://github.com/DjentleViBe/PIRO/wiki/how-to-build) link at the end.

I have added [build](https://github.com/DjentleViBe/PIRO/wiki/how-to-build) link to both https://github.com/DjentleViBe/PIRO/wiki/end-to-end-example and https://github.com/DjentleViBe/PIRO/wiki/Platforms#build-platform. They should all point to the same page now.

Similarly, [Run](https://github.com/DjentleViBe/PIRO/wiki/how-to-run) links are added for the installation sections.
