### 2025-11-27 3c8dabb
make device corrections
