+++ makedevice.sh
@@ -1,3 +1,7 @@
+LOCAL_PREFIX="$(pwd)/local"
+if [[ -d "$LOCAL_PREFIX" ]]; then
+    source ./scripts/linux/env_setup.sh
+fi
 if [[ "$OSTYPE" == "linux-gnu"* ]]; then
     make -f Makefile.devices clean
     make -f Makefile.devices
