### 2025-12-02 8c7deee
removed driver installation
