+++ scripts/linux/Piro_requirements_run_rootful.sh
@@ -82,13 +82,16 @@ if [ -f /etc/debian_version ]; then
         echo "NVIDIA GPU detected, installing NVIDIA OpenCL ICD..."
         DRIVER_VERSION=$(nvidia-smi --query-gpu=driver_version --format=csv,noheader | head -n1)
         DRIVER_MAJOR=$(echo "$DRIVER_VERSION" | cut -d. -f1)
-        apt install -y "libnvidia-compute-${DRIVER_MAJOR}"
+        echo "Please install the latest NVIDIA drivers by running:"
+        echo "apt install -y "libnvidia-compute-575""
     elif lspci | grep -i amd > /dev/null; then
         echo "AMD GPU detected, installing AMD OpenCL..."
-        apt install -y rocm-opencl-runtime
+        echo "Please install the latest AMD drivers by running:"
+        echo "apt install -y rocm-opencl-runtime"
     elif lspci | grep -i intel > /dev/null; then
         echo "INTEL GPU detected, installing INTEL OpenCL..."
-        apt install -y intel-opencl-icd
+        echo "Please install the latest Intel drivers by running:"
+        echo "apt install -y intel-opencl-icd"
     else
         echo "No supported GPU detected for OpenCL."
     fi
@@ -111,14 +114,14 @@ elif [ -f /etc/redhat-release ]; then
         if rpm -q xorg-x11-drv-nvidia &>/dev/null; then
             echo "NVIDIA proprietary driver package is installed."
         else
-            echo "NVIDIA GPU detected, installing NVIDIA OpenCL..."
-            $SUDO dnf install https://download1.rpmfusion.org/free/fedora/rpmfusion-free-release-$(rpm -E %fedora).noarch.rpm
-            $SUDO dnf install https://download1.rpmfusion.org/nonfree/fedora/rpmfusion-nonfree-release-$(rpm -E %fedora).noarch.rpm
-            dnf install -y --allowerasing xorg-x11-drv-nvidia-cuda
+            echo "NVIDIA GPU detected, please install the latest NVIDIA drivers"
+            #$SUDO dnf install https://download1.rpmfusion.org/free/fedora/rpmfusion-free-release-$(rpm -E %fedora).noarch.rpm
+            #$SUDO dnf install https://download1.rpmfusion.org/nonfree/fedora/rpmfusion-nonfree-release-$(rpm -E %fedora).noarch.rpm
+            #dnf install -y --allowerasing xorg-x11-drv-nvidia-cuda
         fi
     elif lspci | grep -i amd > /dev/null; then
-        echo "AMD GPU detected, installing AMD OpenCL..."
-        yum install -y rocm-opencl
+        echo "AMD GPU detected, please insstall AMD OpenCL drivers."
+        # yum install -y rocm-opencl
     fi
 
 elif [ -f /etc/arch-release ]; then
@@ -149,20 +152,21 @@ elif [ -f /etc/arch-release ]; then
         wget -q "$PACKAGE_URL" || { echo "Download failed"; exit 1; }
 
         # Install the package
-        echo "Installing $PACKAGE_NAME..."
-        $SUDO pacman -U --noconfirm "$PACKAGE_NAME" --overwrite usr/lib/libnvidia-opencl.so\* || { echo "Installation failed"; rm -f "$PACKAGE_NAME"; exit 1; }
+        echo "Package is downloaded: Please install by running:"
+        
+        echo "$SUDO pacman -U --noconfirm "$PACKAGE_NAME" --overwrite usr/lib/libnvidia-opencl.so\*"
 
         # Delete the package file after install
-        echo "Deleting package file $PACKAGE_NAME..."
-        rm -f "$PACKAGE_NAME"
[…]
