### 2025-12-02 43f6a77
build device
