+++ .github/workflows/test-build.yaml
@@ -33,6 +33,8 @@ jobs:
           echo "Running rootless on architecture: ${{ matrix.arch }}"
           bash scripts/linux/Piro_requirements_run_rootless.sh
           sudo apt-get install pocl-opencl-icd -y
+      - name: Build device 
+        run: bash ./makedevice.sh
       - name: Setup devices
         run: |
           ./scripts/linux/PIRO_devices_LIN
@@ -53,6 +55,8 @@ jobs:
           echo "Running on architecture: ${{ matrix.arch }}"
           sudo apt-get update
           sudo scripts/linux/Piro_requirements_run_rootful.sh
+      - name: Build device 
+        run: bash ./makedevice.sh
       - name: Setup devices
         run: ./scripts/linux/PIRO_devices_LIN
       - name: Build and test
