### 2025-12-02 DjentleViBe/PIRO#34 @DjentleViBe
> If you are only using `sysroot` to define the root directory for libraries, why can't you just use the typical `include` argument? I'm using a rootful setup during package installation, but I'm confused why I would still need root privileges for the compilation? I **did** manage to compile successfully by removing the `sysroot` flag.
>
> If `sysroot` is used for cross-platform compilation and a container is not possible/preferable, an ideal and clean compilation workflow would be:
>
> 1. Install packages (either by scripts provided or by users on their own)
> 2. Verify the installation
> 3. Set up the compilation and simulation environment
> 4. Verify the environment
> 5. Compile
> 6. Simulate and repeat step 1 if modification is necessary
>
> The verification can either be some tests (which you have created one for step 2 with `check_tool` func), or a configuration file generated by the program, so users can check and compare. I'm pretty much stuck on step 3 and 4, and have to tweak the Makefile to compile, so this is not ideal. Honestly speaking, I don't know if I installed the packages and the environment was setup correctly. Ideally, for a cross-platform compilation, I'd suggest separating system configuration and environment setup steps from the compilation scripts. For example, the scripts could create a human-readable configuration file with compiler and library info (similar to what Rust is doing with `.toml`) and have it verified by human. It will be much easier to debug this way.

The build script is fully rootless. So if all the packages are installed properly inside that folder, there should not be any missing library issue.
Anyway I have [removed](https://github.com/DjentleViBe/PIRO/commit/e9ffbf4d5dc0f4557c0f294672f05b62ed066916) the ```sysroot``` flag. Compiling should use the include argument for dependencies.
```makedevices.sh``` was missing the environment setup step, it is also [added](https://github.com/DjentleViBe/PIRO/actions/runs/19860267545/job/56908279583#step:5:1) to ci/cd.
