### 2025-12-02 DjentleViBe/PIRO#34 @DjentleViBe
> > In file included from /usr/include/c++/15/ext/string_conversions.h:45,
> >                  from /usr/include/c++/15/bits/basic_string.h:4444,
> >                  from /usr/include/c++/15/string:56,
> >                  from /usr/include/c++/15/bits/locale_classes.h:42,
> >                  from /usr/include/c++/15/bits/ios_base.h:43,
> >                  from /usr/include/c++/15/ios:46,
> >                  from /usr/include/c++/15/bits/ostream.h:43,
> >                  from /usr/include/c++/15/ostream:42,
> >                  from /usr/include/c++/15/iostream:43,
> >                  from /PIRO/src/bc.cpp:1:
> > /usr/include/c++/15/cstdlib:83:15: fatal error: stdlib.h: No such file or directory
> >    83 | #include_next <stdlib.h>

So the ubuntu image provided by GitHub CI / CD ships with the standard libraries ```iostream``` etc and so the error is not encountered in testing.
- ```Piro_requirements_run_rootless.sh``` works for program execution.
- ```Piro_requirements_run_rootful.sh``` works for program execution.
- ```Piro_requirements_build.sh```, which is currently rootless, does not work for program compilation without OS supplied standard libraries. #include_next expects stdlib.h to be in a system include directory (like /usr/include) which is solved by using ```--sysroot``` flag when using a local copy.

Do you have a recommendation for installing this with the already provided rootless script (I am not sure its even necessary to make the building process without privileges and without ```sysroot```)? It could be left to the user which is similar to having to ask to follow the [dependencies](https://github.com/DjentleViBe/PIRO/wiki/Dependencies), which I assume the user will do anyway with root.
I believe it's not possible to entirely make it local for program _compiling_ without building gcc from source to ensure ```features.h, stddef.h, stdarg.h, <iostream>``` is available. On the official ubuntu image from Docker Hub (does not contain any c++ libraries), installing ```g++``` from package manager solves everything. Would an approach like this be suitable?
