### 2025-12-16 #8750 @matthewfeickert
@Yurlungur can you please comment on if there are installation issues in https://github.com/DjentleViBe/PIRO/issues/34 that are still a problem for you?

@Brookluo, that's great to hear that you were able to install the software finally. The last few posts in https://github.com/DjentleViBe/PIRO/issues/34 I believe were directed towards you (e.g.https://github.com/DjentleViBe/PIRO/issues/34#issuecomment-3602720466) to get feedback on how you were performing things. Can you please respond to them in https://github.com/DjentleViBe/PIRO/issues/34?
