### 2026-05-01 #8750 @Yurlungur
> @Yurlungur, @Brookluo 👋 (cross-posting between email and here) Can you please comment on if you are still going to be able to engage in the JOSS review? I know that you both had been blocked by https://github.com/DjentleViBe/PIRO/issues/34 from progressing in the review, but I don't know if you have been able to give it another attempt given further discussion. If you are unable to get things installed given the instructions this is useful feedback for the review.

@matthewfeickert I don't recall if I had the time to get past the point where I was blocked in the issue. I was pretty overwhelmed with other obligations when I backed out of the review.
